default https and move func to utils
