+++ cmd/relayer/setup/setup.go
@@ -1,13 +1,10 @@
 package setup
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"net/http"
 	"os"
 	"strings"
-	"time"
 
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
@@ -272,7 +269,7 @@ func getPreRunInfo(home string) (*consts.RollappData, *consts.HubData, string, e
 		return nil, nil, "", err
 	}
 	pterm.Info.Printf("validating rollapp rpc endpoint: %s\n", raRpc)
-	validatedRpc, err := validateRollappRPCEndpoint(raID, raRpc)
+	validatedRpc, err := sequencerutils.ValidateRollappRPCEndpoint(raID, raRpc)
 	if err != nil {
 		pterm.Error.Println("rollapp rpc endpoint validation failed:", err)
 		return nil, nil, "", err
@@ -283,78 +280,6 @@ func getPreRunInfo(home string) (*consts.RollappData, *consts.HubData, string, e
 		RpcUrl: validatedRpc,
 	}
 	return &raData, hd, kb, nil
-}
-
-type tendermintStatusResponse struct {
-	Result struct {
-		NodeInfo struct {
-			Network string `json:"network"`
-		} `json:"node_info"`
-	} `json:"result"`
-}
-
-func validateRollappRPCEndpoint(rollappID, endpoint string) (string, error) {
-	trimmed := strings.TrimSpace(endpoint)
-	if trimmed == "" {
-		return "", fmt.Errorf("rollapp rpc endpoint is empty")
-	}
-
-	normalized := normalizeRpcScheme(trimmed)
-	normalized = strings.TrimRight(normalized, "/")
-
-	statusURL := normalized + "/status"
-	client := &http.Client{
-		Timeout: 10 * time.Second,
-	}
-
-	resp, err := client.Get(statusURL)
-	if err != nil {
-		return "", fmt.Errorf("failed to reach rollapp rpc endpoint %s: %w", normalized, err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf(
-			"rollapp rpc endpoint %s returned %s from /status",
-			normalized,
-			resp.Status,
-		)
-	}
-
-	var status tendermintStatusResponse
-	if err := json.NewDecoder(resp.Body).Decode(&status); err != nil {
-		return "", fmt.Errorf("failed to decode /status from %s: %w", statusURL, err)
-	}
-
-	chainID := strings.TrimSpace(status.Result.NodeInfo.Network)
-	if chainID == "" {
-		return "", fmt.Errorf("rollapp rpc endpoint %s returned empty chain-id in /status response", normalized)
-	}
-	if chainID != rollappID {
-		return "", fmt.Errorf(
-			"rollapp rpc endpoint %s reports chain-id %s, expected %s",
-			normalized,
-			chainID,
-			rollappID,
-		)
-	}
-
-	return normalized, nil
-}
-
-func normalizeRpcScheme(endpoint string) string {
-	switch {
-	case strings.HasPrefix(endpoint, "tcp://"):
-		return "http://" + strings.TrimPrefix(endpoint, "tcp://")
-	case strings.HasPrefix(endpoint, "ws://"):
-		return "http://" + strings.TrimPrefix(endpoint, "ws://")
-	case strings.HasPrefix(endpoint, "wss://"):
-		return "https://" + strings.TrimPrefix(endpoint, "wss://")
-	case strings.HasPrefix(endpoint, "http://"), strings.HasPrefix(endpoint, "https://"):
-		return endpoint
-	default:
-		return "http://" + endpoint
-	}
 }
 
 func installRelayerDependencies(
